# ast: hash containers on insert/update