With this change, we'll calculate hashes as soon as we create or update arrays/sets/objects. This way, we avoid the race detected in #4345 when using PrepareForEval.

❓ Any catastrophic consequences I haven't foreseen?

Fixes #4345.